# Early Perf benchmark for SchemaTransformer