… src/test/resources/large-schema-3.graphqls