Early results on my local machine (Macbook pro 2.4 GHz 8-Core Intel Core i9 32 GB 2667 MHz DDR4)

```
# Warmup: 1 iterations, 10 s each
# Measurement: 1 iterations, 10 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Average time, time/op
# Benchmark: benchmark.SchemaTransformerBenchmark.benchMarkSchemaTransformerRemove

# Run progress: 50.00% complete, ETA 00:02:59
# Fork: 1 of 1
# Warmup Iteration   1: 74264.966 ms/op
Iteration   1: 28894.776 ms/op


Result "benchmark.SchemaTransformerBenchmark.benchMarkSchemaTransformerRemove":
  28894.776 ms/op


# Run complete. Total time: 00:06:02

REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
experiments, perform baseline and negative tests that provide experimental control, make sure
the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
Do not assume the numbers tell you what you want them to tell.

Benchmark                                                    Mode  Cnt      Score   Error  Units
SchemaTransformerBenchmark.benchMarkSchemaTransformerAdd     avgt       30603.825          ms/op
SchemaTransformerBenchmark.benchMarkSchemaTransformerRemove  avgt       28894.776          ms/op
```